Updated the in-file documentation.
Realistically, neither of the concept or reference stylesheets depends
on asciidoc-dita-topic or the modular documentation guidelines.
The only topic type that is sensitive to it is task as generic DITA
topics do not provide any other way how to distinguish context, steps,
prerequisites, results, next steps, or any other sections.