Added a template for topic outputclass.
The asciidoctor-dita-topic plug-in inserts an outputclass to the
generated topic to identify the original content type of the converted
AsciiDoc module. This is so that the correct XSLT to perform the
transformation can be selected programmatically.

Once a generic topic is converted to a DITA task, this outputclass
becomes redundant. I therefore added a template to remove it during
content transformation.