Added an experimental stylesheet for monolithic topics.
DITA is designed to encourage splitting the content in smaller,
specialized topic types rather then holding larger documents in a single
file. That said, the generic DITA topic allows for nesting of topics and
can therefore be used for this purpose. This stylesheet takes large
DITA topics generated with asciidoctor-dita-topic and converts nested
sections to nested topics.